Created base class for builder tests
